proyecto 2 admin/proyecto/services/servicio-1/requirements.txt
fastapi
uvicorn[standard]
psycopg[binary]
passlib[bcrypt]
bcrypt==3.2.0
python-jose[cryptography]
python-multipart
sqlmodel
sqlalchemy
sqlmodel